Fixed bug with relationship event generation.
The contains(x,y) method was given the click point in local coordinates, but was working under the assumption of it's parents
coordinates. Workaround was to add the bounding box position before doing the intersection check. Long term, we need to store
the relationship data in local coordinates and do the conversion only when accessing the ClassNode's positions.